1、newFeature:修改通过编译参数指定平台币精度，同步修改dockerfiles和ReadMe文件 2、bugfix:修改listproducers命令中关于票数根据系统币精度显示 3、optimizationFunction:优化getaccount命令中投票信息的展示 4、bugfix:修改环境依赖脚本中安装kafka插件时未退出目录直接删除目录导致的警告问题

+++ scripts/fscio_build_darwin.sh
@@ -542,6 +542,13 @@
 			exit 1
 		fi
 
+	    if ! cd "${TEMP_DIR}"
+		then
+			printf "\\tUnable to enter directory %s.\\n" "${TEMP_DIR}"
+			printf "\\tExiting now.\\n\\n"
+			exit 1;
+		fi
+		
 		if ! rm -rf "${TEMP_DIR}/librdkafka"
 		then
 			printf "\\tUnable to remove directory %s/librdkafka.\\n" "${TEMP_DIR}"
